requirements.txt
Flask
slackclient